requirements.txt
torch
torchvision
pillow
matplotlib
tqdm
scikit-image
pytorch-msssim